Move operation string conversion to git change class
So that we can pretty-print both git operations as well
as their corresponding operations.
