+++ ueimporter/op.py
@@ -37,18 +37,11 @@ class OpValidation:
 
 
 class Operation:
-    _OP_DESC = ''
-
-    @classmethod
-    @property
-    def op_desc(cls):
-        return cls._OP_DESC
-
     def __init__(self, change):
         self._change = change
 
     def __str__(self):
-        return f'{self.__class__.op_desc} {self.filename}'
+        return str(self._change)
 
     @property
     def filename(self):
@@ -102,20 +95,6 @@ class MoveOp(Operation):
     def target_filename(self):
         return self._change.target_filename
 
-    def __str__(self):
-        common = path_util.commonpath(self.filename, self.target_filename)
-        if common:
-            from_relative = self.filename.relative_to(common)
-            to_relative = self.target_filename.relative_to(common)
-            return \
-                f'Move {from_relative}\n' \
-                f'  to {to_relative}\n' \
-                f'  in {common}'
-        else:
-            return \
-                f'Move {self.filename}\n' \
-                f'  to {self.target_filename}'
-
     def validate(self, source_root, target_root):
         if self.filename == self.target_filename:
             return OpValidation.invalid(
@@ -132,12 +111,3 @@ class MoveOp(Operation):
         if target_root.joinpath(self.target_filename).is_file():
             return OpValidation.invalid_exist(self.target_filename, source_root)
         return OpValidation.valid()
-
-
-# Register operations, and set up descriptions
-_OP_DESC_REGEX = re.compile('^([a-zA-Z]*)Op$')
-_OPERATIONS = [AddOp, DeleteOp, ModifyOp, MoveOp]
-for op_class in _OPERATIONS:
-    match = _OP_DESC_REGEX.match(op_class.__name__)
-    assert match
-    op_class._OP_DESC = match.group(1)
